DO NOT USE - trying to add a different type of s-curve

+++ Utils/RD53Shared.h
@@ -59,9 +59,9 @@ T mtxInversion(const boost::numeric::ublas::matrix<T>& input, boost::numeric::ub
     if((isSingular = boost::numeric::ublas::lu_factorize(mLU, pivots)) != false) return 0;
 
     T det = 1;
-    for(auto i = 0; i < pivots.size(); i++)
+    for(auto i = 0; i < (int)pivots.size(); i++)
     {
-        if(pivots(i) != i) det *= -1;
+        if((int)pivots(i) != i) det *= -1;
         det *= mLU(i, i);
     }
